script: Move CustomTraceable to script_bindings.
